requirements.txt
transformers
torch
gradio
Pillow
accelerate
huggingface_hub